add Auth plugin
