+++ package.json
@@ -13,7 +13,8 @@
     "@fastify/helmet": "^8.0.1",
     "@fastify/rate-limit": "^6.0.0",
     "axios": "^0.27.2",
-    "fastify": "^3.29.0"
+    "fastify": "^3.29.0",
+    "fastify-plugin": "^3.0.1"
   },
   "devDependencies": {
     "nodemon": "^2.0.16"
